tests/version: Move tests into `routes` submodules
