+++ src/tests/routes/crates/mod.rs
@@ -1,2 +1,4 @@
 pub mod downloads;
+mod new;
+mod read;
 pub mod versions;
