+++ src/tests/routes/crates/versions/mod.rs
@@ -1,2 +1,4 @@
+mod authors;
 pub mod dependencies;
 pub mod download;
+mod read;
