+++ src/tests/routes/mod.rs
@@ -14,3 +14,4 @@
 pub mod categories;
 pub mod category_slugs;
 pub mod crates;
+pub mod versions;
