+++ src/tests/routes/versions/mod.rs
@@ -1,0 +1,2 @@
+pub mod list;
+pub mod read;
